completion of aavegotchi-agip-37-unstaked-lp strategy
