--- src/strategies/aavegotchi-agip-37-smart-contract/README.md
@@ -1,21 +1,0 @@
-# Aavegotchi AGIP 37
-
-## Description
-
-This snapshot strategy enables voting power for the following assets as a part if Aavegotchi AGIP-37
-
-- amGHST
-- wapGHST
-- wapGHST staked for GLTR
-- GHST-FUD LP staked for GLTR
-- GHST-FOMO LP staked for GLTR
-- GHST-ALPHA LP staked for GLTR
-- GHST-KEK LP staked for GLTR
-- GHST-GLTR LP staked for GLTR
-- GHST-USDC LP staked for GLTR
-- GHST-WMATIC LP staked for GLTR
-- Unstaked GHST-FUD, GHST-FOMO, GHST-ALPHA, GHST-KEK, GHST-GLTR LP tokens
-
-## References
-
-Aavegotchi AGIP 37: https://snapshot.org/#/aavegotchi.eth/proposal/0x9923aab6825158ec2503d88e3ee2f9c5fbb12000581d06343ac9829aa59b66a6
