--- src/strategies/aavegotchi-agip-37-smart-contract/examples.json
@@ -1,14 +1,0 @@
-[
-  {
-    "name": "AGIP 37: Uses a smart contract to calculating Aavegotchi AGIP-37 voting power.",
-    "strategy": {
-      "name": "aavegotchi-agip-37-smart-contract",
-      "params": {
-        "votingPowerAddress": "0x66054Cd5b64fc1Fc0330DD711a7956e468f97933"
-      }
-    },
-    "network": "137",
-    "addresses": ["0x26cf02F892B04aF4Cf350539CE2C77FCF79Ec172", "0x027Ffd3c119567e85998f4E6B9c3d83D5702660c", "0xc4cb6cb969e8b4e309ab98e4da51b77887afad96"],
-    "snapshot": 33770634
-  }
-]
