+++ src/strategies/aavegotchi-agip-37-unstaked-lp/README.md
@@ -1,0 +1,11 @@
+# Aavegotchi AGIP 37 - Unstaked LP Pairs
+
+## Description
+
+This snapshot strategy enables voting power for the following assets as a part if Aavegotchi AGIP-37
+
+- Unstaked GHST-FUD, GHST-FOMO, GHST-ALPHA, GHST-KEK, and GHST-GLTR LP tokens
+
+## References
+
+Aavegotchi AGIP 37: https://snapshot.org/#/aavegotchi.eth/proposal/0x9923aab6825158ec2503d88e3ee2f9c5fbb12000581d06343ac9829aa59b66a6
