+++ src/strategies/aavegotchi-agip-37-unstaked-lp/examples.json
@@ -1,0 +1,14 @@
+[
+  {
+    "name": "AGIP 37: Uses a smart contract to calculate the voting power for unstaked GHST-FUD, GHST-FOMO, GHST-ALPHA, GHST-KEK, and GHST-GLTR LP tokens",
+    "strategy": {
+      "name": "aavegotchi-agip-37-unstaked-lp",
+      "params": {
+        "votingPowerAddress": "0x0D00800489dcAb402D4A17C5BAAfe80c4E22a5d9"
+      }
+    },
+    "network": "137",
+    "addresses": ["0x26cf02F892B04aF4Cf350539CE2C77FCF79Ec172", "0x027Ffd3c119567e85998f4E6B9c3d83D5702660c", "0xc4cb6cb969e8b4e309ab98e4da51b77887afad96"],
+    "snapshot": 33861855
+  }
+]
